+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [2.0.4] - 2019/10/01
 
-* StoreTester: waitError and waitErrorGetLast.
+* StoreTester: waitUntilError and waitUntilErrorGetLast.
 
 ## [2.0.3] - 2019/09/21